Div/appeals 46136 (#21781)
* Created new file for correspondence and modified some changes from correspondence_auto_assign to user

* Shifted the auto_assign_permissions method to users.rb file

* Nod Permissions method is shifted to users

* Modified the create! method with find_or_create_by

* Fix for inbound_ops_team_superuser check

* Moved correspondence_types and package_document_types into new seed file

* Move code from queue_correspondence to new seed

* Added a document type for every correspondence

* Removed all the related correspondences seed files

* Deleted the merged files again

* Modified the rake tasks file and renamed the file to correspondence

* Modified the appeals creation for veteran from 10 to 15

* Updated the spec files with seed class

* Users sepec count update
