+++ db/seeds/users.rb
@@ -59,6 +59,11 @@ module Seeds
       create_oai_team_user
       create_occ_team_user
       create_inbound_ops_team_user
+      create_inbound_ops_team_auto_assign_user
+      create_inbound_ops_team_user_with_no_permissions
+      create_inbound_ops_team_supervisor
+      create_inbound_ops_team_superuser
+      create_inbound_ops_team_nod_user
       create_cavc_lit_support_user
       create_pulac_cerullo_user
       create_mail_team_user
@@ -367,12 +372,12 @@ module Seeds
       AodTeam.singleton.add_user(u)
 
       root_task = create(:root_task)
-      mail_task = ::AodMotionMailTask.create!(
+      mail_task = ::AodMotionMailTask.find_or_create_by(
         appeal: root_task.appeal,
         parent_id: root_task.id,
         assigned_to: MailTeam.singleton
       )
-      ::AodMotionMailTask.create!(
+      ::AodMotionMailTask.find_or_create_by(
         appeal: root_task.appeal,
         parent_id: mail_task.id,
         assigned_to: AodTeam.singleton
@@ -404,14 +409,160 @@ module Seeds
       OaiTeam.singleton.add_user(u)
     end
 
-    def create_inbound_ops_team_user
-      u = User.create!(station_id: 101, css_id: "INBOUND_OPS_TEAM_ADMIN_USER", full_name: "Jon MailTeam Snow Admin")
-      InboundOpsTeam.singleton.add_user(u)
-      OrganizationsUser.make_user_admin(u, InboundOpsTeam.singleton)
+    def create_auto_assign_permissions
+      OrganizationPermission.valid_permission_names.each do |permission|
+        OrganizationPermission.find_or_create_by(
+          permission: permission,
+          organization: InboundOpsTeam.singleton
+        ) do |perm|
+          perm.enabled = true
+          perm.description = Faker::Hipster.sentence
+        end
+      end
+      OrganizationPermission.find_by(permission: "superuser").update!(
+        description: "Superuser: Split, Merge, and Reassign",
+        default_for_admin: true,
+      )
+      OrganizationPermission.find_by(permission: "auto_assign").update!(description: "Auto-Assignment")
+      OrganizationPermission.find_by(permission: "receive_nod_mail").update!(
+        description: "Receieve \"NOD Mail\"",
+        parent_permission: OrganizationPermission.find_by(permission: "auto_assign")
+      )
+    end
 
-      u = User.create!(station_id: 101, css_id: "INBOUND_OPS_TEAM_MAIL_INTAKE_USER",
-        full_name: "Jon MailTeam Snow Mail Intake", roles: ["Mail Intake"])
-      InboundOpsTeam.singleton.add_user(u)
+    def create_inbound_ops_team_user
+      # create the necessary permissions first
+      create_auto_assign_permissions
+      # Admin user
+      admin_user_info = [
+       { css_id: "INBOUND_OPS_TEAM_ADMIN_USER", full_name: "Jon MailTeam Snow Admin", admin: true }
+      ]
+      # Non-admin users
+      non_admin_users_info = [
+        { css_id: "INBOUND_OPS_TEAM_MAIL_INTAKE_USER", full_name: "Jon MailTeam Snow Mail Intake", roles: ["Mail Intake"] },
+        { css_id: "INBOUND_OPS_TEAM_USER_U1", full_name: "Cedar Rain", roles: ["Mail Intake"] }
+      ]
+
+      users_info = admin_user_info + non_admin_users_info
+      # Create users
+      users_info.each do |user_info|
+        new_user = User.find_or_create_by!(
+          station_id: 101,
+          css_id: user_info[:css_id],
+          full_name: user_info[:full_name],
+          roles: user_info[:roles] || []
+        )
+        InboundOpsTeam.singleton.add_user(new_user)
+        OrganizationsUser.make_user_admin(new_user, InboundOpsTeam.singleton) if user_info[:admin]
+      end
+    end
+
+    def create_inbound_ops_team_auto_assign_user
+      create_auto_assign_permissions
+      users_info = [
+        { css_id: "INBOUND_OPS_TEAM_MAIL_INTAKE_USER_AUTO_ASSIGN_A1", full_name: "Ember Sky" },
+        { css_id: "INBOUND_OPS_TEAM_MAIL_INTAKE_USER_AUTO_ASSIGN_A2", full_name: "Aspen Ridge" }
+      ]
+      users_info.map do |user_info|
+        new_user = create_user(user_info)
+        org_user = OrganizationsUser.find_or_create_by!(organization: InboundOpsTeam.singleton, user: new_user)
+        auto_assign = OrganizationPermission.find_by(organization: InboundOpsTeam.singleton, permission: "auto_assign")
+        OrganizationUserPermission.find_or_create_by!(
+          organization_permission: auto_assign,
+          organizations_user: org_user
+        ) do |op|
+          op.permitted = true
+        end
+      end
+    end
+
+    def create_inbound_ops_team_user_with_no_permissions
+      create_auto_assign_permissions
+      users_info = [
+        { css_id: "INBOUND_OPS_TEAM_MAIL_INTAKE_USER_NP1", full_name: "Noah Taylor" },
+        { css_id: "INBOUND_OPS_TEAM_MAIL_INTAKE_USER_NP2", full_name: "Emma Brown" }
+      ]
+      users_info.map do |user_info|
+        new_user = User.find_or_create_by!(
+          station_id: 101,
+          css_id: user_info[:css_id],
+          full_name: user_info[:full_name],
+          roles: ["Mail Intake"]
+        )
+        OrganizationsUser.find_or_create_by!(organization: InboundOpsTeam.singleton, user: new_user)
+      end
+    end
+
+    def create_inbound_ops_team_supervisor
+      create_auto_assign_permissions
+      users_info = [
+        { css_id: "INBOUND_OPS_TEAM_ADMIN_USER_S1", full_name: "Caleb Mitchell" },
+        { css_id: "INBOUND_OPS_TEAM_ADMIN_USER_S2", full_name: "Scarlett Reed" }
+      ]
+      users_info.map do |user_info|
+        user = User.find_or_create_by!(
+          station_id: 101,
+          css_id: user_info[:css_id],
+          full_name: user_info[:full_name],
+          roles: ["Mail Intake"]
+        )
+        InboundOpsTeam.singleton.add_user(user)
+        OrganizationsUser.make_user_admin(user, InboundOpsTeam.singleton)
+      end
+    end
+
+    def create_inbound_ops_team_superuser
+      create_auto_assign_permissions
+      users_info = [
+        { css_id: "INBOUND_OPS_TEAM_SUPERUSER1", full_name: "Willow Green" },
+        { css_id: "INBOUND_OPS_TEAM_SUPERUSER2", full_name: "Jasper Bloom" }
+      ]
+      users_info.map do |user_info|
+        new_user = User.find_or_create_by!(
+          station_id: 101,
+          css_id: user_info[:css_id],
+          full_name: user_info[:full_name],
+          roles: ["Mail Intake"]
+        )
+        InboundOpsTeam.singleton.add_user(new_user)
+        superuser_permission = OrganizationPermission.find_by(permission: 'superuser', organization_id: InboundOpsTeam.singleton.id)
+        OrganizationUserPermission.find_or_create_by!(
+          organization_permission: superuser_permission,
+          organizations_user: OrganizationsUser.find_by(user_id: new_user.id),
+          permitted: true
+        )
+      end
+    end
+
+    def create_inbound_ops_team_nod_user
+      create_auto_assign_permissions
+      users_info = [
+        { css_id: "INBOUND_OPS_TEAM_MAIL_INTAKE_USER_NOD1", full_name: "Alexandr Johnson" },
+        { css_id: "INBOUND_OPS_TEAM_MAIL_INTAKE_USER_NOD2", full_name: "Sopia Williams" }
+      ]
+      users_info.map do |user_info|
+        new_user = create_user(user_info)
+        org_user = OrganizationsUser.find_or_create_by!(organization: InboundOpsTeam.singleton, user: new_user)
+        receive_nod_mail = OrganizationPermission.find_by(
+          organization: InboundOpsTeam.singleton,
+          permission: "receive_nod_mail"
+        )
+        OrganizationUserPermission.find_or_create_by!(
+          organization_permission: receive_nod_mail,
+          organizations_user: org_user
+        ) do |op|
+          op.permitted = true
+        end
+      end
+    end
+
+    def create_user(user_info)
+      User.find_or_create_by!(
+        station_id: 101,
+        css_id: user_info[:css_id],
+        full_name: user_info[:full_name],
+        roles: ["Mail Intake"]
+      )
     end
 
     def create_cavc_lit_support_user
@@ -530,6 +681,7 @@ module Seeds
                           roles: ["Mail Intake"])
       CDAControlGroup.singleton.add_user(leo)
       OrganizationsUser.make_user_admin(leo, CDAControlGroup.singleton)
+
       casey = User.find_or_create_by!(station_id: 101,
                             css_id: "CDAUSERCASEY",
                             full_name: "Casey CDAC_User Jones",
